separate a bit
